# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MoveMove API",
    "description": "<h3>MoveMove 募募 募資網站 API 列表</h3><li>您可以在此查看所有的 API 與線上測試</li><li>若有問題反饋可以到我們的<a href=\"https://github.com/Move-Move-CrowdFunding/movemove_api\">Github</a>發issue</li>",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "in": "headers",
      "name": "Authorization",
      "description": "請填寫 token"
    }
  },
  "paths": {
...
    "/project/": {
      "get": {
        "tags": [
          "Projects - 提案"
        ],
        "description": "提案列表",
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "當前第幾頁",
            "type": "number",
            "default": "1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "一頁有幾筆",
            "type": "number",
            "default": "10"
          },
          {
            "name": "categoryKey",
            "in": "query",
            "description": "提案分類: 0-全部 1-教育 2-弱勢救助 3-國際支援 4-兒少福利 5-長者 6-婦女",
            "type": "number",
            "default": "0"
          },
          {
            "name": "isExpired",
            "in": "query",
            "description": "是否包含已結束募資的提案",
            "type": "boolean",
            "default": "false"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "頁面排序: 1-由新到舊 2-由舊到新",
            "type": "number",
            "default": "1"
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "關鍵字查詢",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "取得募資列表成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "取得募資列表成功"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "663a5750e34b6703e22a9f60"
                      },
                      "introduce": {
                        "type": "string",
                        "example": "專業金援團隊，弱勢族群救星，幫助許多需要協助的家庭。"
                      },
                      "teamName": {
                        "type": "string",
                        "example": "弱勢救星"
                      },
                      "email": {
                        "type": "string",
                        "example": "nomail@mail.com"
                      },
                      "phone": {
                        "type": "string",
                        "example": "0938938438"
                      },
                      "title": {
                        "type": "string",
                        "example": "第二筆提案"
                      },
                      "categoryKey": {
                        "type": "number",
                        "example": 2
                      },
                      "targetMoney": {
                        "type": "number",
                        "example": 50000
                      },
                      "startDate": {
                        "type": "number",
                        "example": 1718121600
                      },
                      "endDate": {
                        "type": "number",
                        "example": 1718985600
                      },
                      "describe": {
                        "type": "string",
                        "example": "一場無情的大火吞噬了整個社區，請幫助無家可歸的民眾。"
                      },
                      "coverUrl": {
                        "type": "string",
                        "example": "https://fakeimg.pl/300/"
                      },
                      "content": {
                        "type": "string",
                        "example": "<p>test</p>"
                      },
                      "videoUrl": {
                        "type": "string",
                        "example": ""
                      },
                      "relatedUrl": {
                        "type": "string",
                        "example": ""
                      },
                      "feedbackItem": {
                        "type": "string",
                        "example": "限量精美小熊維尼"
                      },
                      "feedbackUrl": {
                        "type": "string",
                        "example": "https://fakeimg.pl/300/"
                      },
                      "feedbackMoney": {
                        "type": "number",
                        "example": 100
                      },
                      "feedbackDate": {
                        "type": "number",
                        "example": 1682524800
                      },
                      "achievedMoney": {
                        "type": "number",
                        "example": 0
                      },
                      "trackingStatus": {
                        "type": "boolean",
                        "example": false
                      }
                    }
                  }
                },
                "pagination": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number",
                      "example": 1
                    },
                    "pageNo": {
                      "type": "number",
                      "example": 1
                    },
                    "pageSize": {
                      "type": "number",
                      "example": 10
                    },
                    "hasPre": {
                      "type": "boolean",
                      "example": false
                    },
                    "hasNext": {
                      "type": "boolean",
                      "example": false
                    },
                    "totalPage": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "取得列表失敗",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "message": {
                  "type": "string",
                  "example": "無此分頁"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects - 提案"
        ],
        "description": "發起提案",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "發起提案格式",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string",
                  "example": "66378c5f9fb5b5e7c300e95c"
                },
                "introduce": {
                  "type": "string",
                  "example": "團隊介紹"
                },
                "teamName": {
                  "type": "string",
                  "example": "提案人姓名/團隊名稱"
                },
                "title": {
                  "type": "string",
                  "example": "提案標題"
                },
                "email": {
                  "type": "string",
                  "example": "聯絡信箱"
                },
                "categoryKey": {
                  "type": "number",
                  "example": 1
                },
                "phone": {
                  "type": "string",
                  "example": "連絡電話"
                },
                "targetMoney": {
                  "type": "number",
                  "example": 50000
                },
                "content": {
                  "type": "string",
                  "example": "<p>提案內容</p>"
                },
                "coverUrl": {
                  "type": "string",
                  "example": "https://fakeimg.pl/300/"
                },
                "describe": {
                  "type": "string",
                  "example": "提案簡介"
                },
                "videoUrl": {
                  "type": "string",
                  "example": "https://youtube.com/"
                },
                "startDate": {
                  "type": "number",
                  "example": 1678982400
                },
                "endDate": {
                  "type": "number",
                  "example": 1679846400
                },
                "relatedUrl": {
                  "type": "string",
                  "example": "https://www.google.com/"
                },
                "feedbackItem": {
                  "type": "string",
                  "example": "限量精美小熊維尼"
                },
                "feedbackUrl": {
                  "type": "string",
                  "example": "https://fakeimg.pl/300/"
                },
                "feedbackMoney": {
                  "type": "number",
                  "example": 100
                },
                "feedbackDate": {
                  "type": "number",
                  "example": 1682524800
                }
              },
              "required": [
                "userId",
                "teamName",
                "title",
                "email",
                "categoryKey",
                "phone",
                "targetMoney",
                "content",
                "coverUrl",
                "describe",
                "startDate",
                "endDate"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "發起提案成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "資料新增成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "發起提案失敗",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "message": {
                  "type": "string",
                  "example": "請輸入提案標題"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/project/{id}": {
      "get": {
        "tags": [
          "Projects - 提案"
        ],
        "description": "單一提案內容",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "提案 id"
          }
        ],
        "responses": {
          "200": {
            "description": "單一提案獲取成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "取得募資列表成功"
                },
                "results": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "663e3f9652027aeb86960016"
                    },
                    "userId": {
                      "type": "string",
                      "example": "663e3f3452027aeb8696000c"
                    },
                    "categoryKey": {
                      "type": "number",
                      "example": 2
                    },
                    "coverUrl": {
                      "type": "string",
                      "example": "https://fakeimg.pl/300/"
                    },
                    "describe": {
                      "type": "string",
                      "example": "一場無情的大火吞噬了整個社區，請幫助無家可歸的民眾。"
                    },
                    "endDate": {
                      "type": "number",
                      "example": 1721577600
                    },
                    "title": {
                      "type": "string",
                      "example": "2"
                    },
                    "videoUrl": {
                      "type": "string",
                      "example": ""
                    },
                    "startDate": {
                      "type": "number",
                      "example": 1720713600
                    },
                    "feedbackItem": {
                      "type": "string",
                      "example": "限量精美小熊維尼"
                    },
                    "feedbackUrl": {
                      "type": "string",
                      "example": "https://fakeimg.pl/300/"
                    },
                    "feedbackMoney": {
                      "type": "number",
                      "example": 100
                    },
                    "feedbackDate": {
                      "type": "number",
                      "example": 1721577600
                    },
                    "targetMoney": {
                      "type": "number",
                      "example": 30000
                    },
                    "teamName": {
                      "type": "string",
                      "example": "弱勢救星"
                    },
                    "phone": {
                      "type": "string",
                      "example": "0938938438"
                    },
                    "email": {
                      "type": "string",
                      "example": "nomail@mail.com"
                    },
                    "relatedUrl": {
                      "type": "string",
                      "example": ""
                    },
                    "introduce": {
                      "type": "string",
                      "example": "專業金援團隊，弱勢族群救星，幫助許多需要協助的家庭。"
                    },
                    "content": {
                      "type": "string",
                      "example": "<p>test</p>"
                    },
                    "achievedMoney": {
                      "type": "number",
                      "example": 2200
                    },
                    "supportCount": {
                      "type": "number",
                      "example": 2
                    },
                    "trackingStatus": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "取得提案失敗",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "message": {
                  "type": "string",
                  "example": "查無此提案"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}